Fix tar invocation (flags)
Not all tar implementations detect (or assume) stdin as their default
read source (ex. BSDs, Solaris, etc.).  In fact, most assume a tape
device (ex. /dev/sa0), so "-f -" must be used.

The author prefers non-hyphenated flags, so for this to work properly,
the flags must be in the correct order, ex. "zxf-".